docs: Remove Japanese documents
We have maintained both English and Japanese documents since its
release. However, we found that it's hard to keep updated both documents
in the view of maintainance cost. We have to focus on the development,
not documentation because this is a research and development project.
This commit removes all Japanese documents and the references.

Signed-off-by: Akira Moroo <akiram@ricsec.co.jp>